Rename clinet to clientconfig package
