+++ pkg/pipelines/clientconfig/client_config.go
@@ -1,4 +1,4 @@
-package client
+package clientconfig
 
 import (
 	"k8s.io/client-go/rest"
